fix taskcotroller

+++ app/Http/Controllers/TasksController.php
@@ -100,8 +100,11 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
+        
         $task = Task::find($id);
-
+         if ((\Auth::user()->id) != $task->user_id){
+            return redirect('/');
+        }
         return view('tasks.edit', [
             'task' => $task,
         ]);
@@ -138,6 +141,9 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
+        if ((\Auth::user()->id) != $task->user_id){
+            return redirect('/');
+        }
         $task->delete();
 
         return redirect('/');
